Move AnalysisType to constants
