+++ cg_hermes/constants/workflow.py
@@ -17,3 +17,10 @@ class Workflow(StrEnum):
     MUTANT: str = "mutant"
     SPRING: str = "spring"
     TAXPROFILER: str = "taxprofiler"
+
+
+class AnalysisType(StrEnum):
+    TUMOR_WGS: str = "tumor_wgs"
+    TUMOR_NORMAL_WGS: str = "tumor_normal_wgs"
+    TUMOR_PANEL: str = "tumor_panel"
+    TUMOR_NORMAL_PANEL: str = "tumor_normal_panel"
