+++ cg_hermes/deliverables.py
@@ -1,4 +1,5 @@
 """Class to represent deliverables file."""
+
 import copy
 import logging
 from typing import FrozenSet
@@ -27,8 +28,7 @@ from cg_hermes.config.mip_rna import MIP_RNA_TAGS
 from cg_hermes.config.mutant import MUTANT_COMMON_TAGS
 from cg_hermes.config.rnafusion import RNAFUSION_TAGS
 from cg_hermes.config.taxprofiler import TAXPROFILER_TAGS
-from cg_hermes.config.workflows import AnalysisType
-from cg_hermes.constants.workflow import Workflow
+from cg_hermes.constants.workflow import AnalysisType, Workflow
 from cg_hermes.exceptions import MissingFileError
 from cg_hermes.models.tags import CGTag, TagMap
 from cg_hermes.models.workflow_deliverables import (
@@ -175,23 +175,23 @@ class Deliverables:
         tag_set = []
         if self.workflow == Workflow.BALSAMIC:
             BALSAMIC_COMMON_TAGS = BALSAMIC_TAGS
-            if self.analysis_type == AnalysisType.tumor_wgs:
+            if self.analysis_type == AnalysisType.TUMOR_WGS:
                 tag_set = TUMOR_ONLY_WGS_TAGS
-            elif self.analysis_type == AnalysisType.tumor_normal_wgs:
+            elif self.analysis_type == AnalysisType.TUMOR_NORMAL_WGS:
                 tag_set = TUMOR_NORMAL_WGS_TAGS
-            elif self.analysis_type == AnalysisType.tumor_panel:
+            elif self.analysis_type == AnalysisType.TUMOR_PANEL:
                 tag_set = TUMOR_ONLY_PANEL_TAGS
             else:
                 tag_set = TUMOR_NORMAL_PANEL_TAGS
         elif self.workflow == Workflow.BALSAMIC_QC:
             BALSAMIC_COMMON_TAGS = BALSAMIC_QC_TAGS
-            if self.analysis_type == AnalysisType.tumor_normal_wgs:
+            if self.analysis_type == AnalysisType.TUMOR_NORMAL_WGS:
                 tag_set = QC_TUMOR_NORMAL_WGS_TAGS
-            elif self.analysis_type == AnalysisType.tumor_normal_panel:
+            elif self.analysis_type == AnalysisType.TUMOR_NORMAL_PANEL:
                 tag_set = QC_TUMOR_NORMAL_PANEL_TAGS
         elif self.workflow == Workflow.BALSAMIC_UMI:
             BALSAMIC_COMMON_TAGS = BALSAMIC_UMI_TAGS
-            if self.analysis_type == AnalysisType.tumor_panel:
+            if self.analysis_type == AnalysisType.TUMOR_PANEL:
                 tag_set = UMI_TUMOR_ONLY_PANEL_TAGS
             else:
                 tag_set = UMI_TUMOR_NORMAL_PANEL_TAGS
